+++ test/test_meter.cc
@@ -16,7 +16,7 @@ using namespace medida;
 TEST(MeterTest, aBlankMeter) {
   Meter meter {"things"};
   EXPECT_EQ("things", meter.event_type());
-  EXPECT_EQ(0, meter.count());
+  EXPECT_EQ(static_cast<std::uint64_t>(0), meter.count());
   EXPECT_NEAR(0.0, meter.mean_rate(), 0.001);
 }
 
@@ -24,7 +24,7 @@ TEST(MeterTest, aBlankMeter) {
 TEST(MeterTest, createFromRegistry) {
   MetricsRegistry registry {};
   auto& meter = registry.NewMeter({"a", "b", "c"}, "things");
-  EXPECT_EQ(0, meter.count());
+  EXPECT_EQ(static_cast<std::uint64_t>(0), meter.count());
   EXPECT_EQ("things", meter.event_type());
 }
 
@@ -32,7 +32,7 @@ TEST(MeterTest, createFromRegistry) {
 TEST(MeterTest, aMeterWithThreeEvents) {
   Meter meter {"things"};
   meter.Mark(3);
-  EXPECT_EQ(3, meter.count());
+  EXPECT_EQ(static_cast<std::uint64_t>(3), meter.count());
 }
 
 
@@ -42,7 +42,7 @@ TEST(MeterTest, meterTiming) {
     meter.Mark();
     std::this_thread::sleep_for(std::chrono::milliseconds(100));
   }
-  EXPECT_EQ(10, meter.count());
+  EXPECT_EQ(static_cast<std::uint64_t>(10), meter.count());
   EXPECT_NEAR(10, meter.mean_rate(), 0.1);
 }
